feat: Remove generic assemble endpoint and move to dsyms
